amazon/rt-services/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
numpy>=1.24
opencv-python-headless>=4.8
pillow>=9.0
torch
